Allow to have a specific implementation of the profiler per OS
